Moving transgenic_enhancer_experiment.py types to its own file like all the other schemas
